requirements.txt
appdirs==1.4.3
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
black==18.9b0
choicesenum==0.4.0
Click==7.0
Django==2.1.3
django-debug-toolbar==1.10.1
django-phonenumber-field==2.1.0
djangorestframework==3.9.0
more-itertools==4.3.0
phonenumberslite==8.10.0
pluggy==0.8.0
py==1.7.0
pytest==4.0.0
pytest-django==3.4.4
pytz==2018.7
sqlparse==0.2.4
toml==0.10.0
